% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brink-data.R
\docType{data}
\name{brink}
\alias{brink}
\title{Small taper dataset from Brink and von Gadow (1986).}
\format{
A data frame with 108 rows and 5 variables:
\describe{
  \item{Tree}{tree identification number (factor)}
  \item{h}{height level, meters}
  \item{dib}{diameter inside bark, centimeters}
  \item{Dob}{dbh outside bark, cm (breast height is 1.35 m)}
  \item{H}{total height, meters}
}
}
\source{
{Brink, C. and von Gadow, K. (1986) "On the use of growth and decay
  functions for modelling stem profiles". \emph{EDV in Medizin und Biologi
  17}, 20-27}
}
\usage{
brink
}
\description{
Taper measurements on 10 \emph{Eucalyptus cloeziana} trees.
}
\keyword{datasets}
