% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decay.R
\name{Id}
\alias{Id}
\title{Integral of decay function.}
\usage{
Id(x, p)
}
\arguments{
\item{x}{Input value(s), possibly a vector.}

\item{p}{Parameter.}
}
\value{
Integral of the decay function between 0 and x.
}
\description{
Integral of decay function.
}
\examples{
   Id(2, 0)
   Id(1.5,0.5)
   Id(2.5, 0.5)
   Id(2.5, -0.5)
   for(p in seq(1, -1, -0.5)) curve(Id(x, p), 0, 3, add=(p != 1))

}
