% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taper.R
\name{unscaled}
\alias{unscaled}
\title{Unscaled curve of cross-sectional area vs height level.}
\usage{
unscaled(h, H, b)
}
\arguments{
\item{h}{Height level(s), possibly a vector.}

\item{H}{Tree total height.}

\item{b}{Vector with the 5 parameters.}
}
\value{
Value(s) proportional to the area(s) at the level(s) \code{h}.
}
\description{
Unscaled curve of cross-sectional area vs height level.
}
\examples{
    unscaled(16, 32, c(2.569, 0, 1.042, 0.3012, -1))

}
