% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyntaper-package.R
\docType{package}
\name{dyntaper-package}
\alias{dyntaper}
\alias{dyntaper-package}
\title{dyntaper: Using a dynamic tree stem profile model}
\description{
Implements the model from García, O. (2015) "Dynamic modelling of tree form", MCFNS 7(1): 9-15.
}
\details{
The main functions relevant to the user are
\describe{
  \item{\code{\link{taper}}:}{Calculates diameters or cross-sectional areas
    for given height levels.}
  \item{\code{\link{hlevel}}:}{Calculates the height level for a given
    diameter or cross-sectional area.}
  \item{\code{\link{volume}}:}{Calculates the volume between two given height
    levels.}
}
The package includes a small example dataset, \code{\link{brink}}.

For a tutorial, see the vignette: \code{vignette("dyntaper", package="dyntaper")}.
}
\references{
García, O. (2015) "Dynamic modelling of tree form". \emph{Mathematical
  and Computational Forestry & Natural-Resource Sciences (MCFNS) 7}(1), 9–15.
  (\url{http://mcfns.net/index.php/Journal/article/view/MCFNS7.1_2}).
Koirala, A., Montes, C. R., Bullock, B. P. and Wagle, B. H. (2021) "Developing
  taper equations for planted teak (\emph{Tectona grandis} L. f.) trees of
  central lowland Nepal". \emph{Trees, Forests and People 5}.
  (\url{https://doi.org/10.1016/j.tfp.2021.100103}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ogarciav/dyntaper}
  \item Report bugs at \url{https://github.com/ogarciav/dyntaper/issues}
}

}
\keyword{internal}
